Initialize the initial framework for the Todo List
Set up the overall structure of the project, including the webpack
configuration, the project files, and the initial project files.
Settled the layout of the page with CSS. Added the add project
functionality to the website, which allows users to add projects.
Also enabled users to toggle between differen projects.